+++ server.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import mongoose from 'mongoose';
 const server = express();
-import ApiRouter from './api/registration.js';
+import ApiRouter from './api/api.js';
 import passportRoutes from './passportConfig';
 
 const bodyParser = require('body-parser');
